[test] : 서비스 분할에 따른 단위테스트 코드 변경
